move things to constants file
