+++ src/views/menus/AccountMenu.tsx
@@ -17,7 +17,7 @@ import {
 } from '@/constants/localization';
 import { isDev } from '@/constants/networks';
 import { SMALL_USD_DECIMALS, USD_DECIMALS } from '@/constants/numbers';
-import { DydxChainAsset, wallets } from '@/constants/wallets';
+import { DydxChainAsset, wallets, WalletType } from '@/constants/wallets';
 
 import { useAccountBalance } from '@/hooks/useAccountBalance';
 import { useAccounts } from '@/hooks/useAccounts';
@@ -52,7 +52,6 @@ import { openDialog } from '@/state/dialogs';
 import { isTruthy } from '@/lib/isTruthy';
 import { MustBigNumber } from '@/lib/numbers';
 import { truncateAddress } from '@/lib/wallet';
-import { WalletType } from '@/lib/wallet/types';
 
 import { MobileDownloadLinks } from '../MobileDownloadLinks';
